Door added, event to open the door left to do

+++ includes/render.h
@@ -17,19 +17,19 @@
 # define RIGHT 2
 # define SQUARE_SIZE 5
 
-void	render(t_texture *config);
-int		destroy_win(t_mlx *mlx);
-void	handle_key(void *param);
-void	mouse_hook_camera(void *data);
-void	raycasting(t_mlx *mlx);
-double	ft_abs(double nb);
-size_t	get_color_from_img(mlx_image_t *img, int x, int y);
-void	texture_to_tab(t_texture *config, t_mlx *mlx);
-void	animation_sprite(t_texture *config, t_mlx *mlx);
-void	put_sprite_to_img(t_mlx *mlx, int x, t_dda *dda, t_raycast *info);
-void	put_sprite_to_img_animation(t_mlx *mlx, int x, t_dda *dda,
+void		render(t_texture *config);
+int			destroy_win(t_mlx *mlx);
+void		handle_key(void *param);
+void		mouse_hook_camera(void *data);
+void		raycasting(t_mlx *mlx);
+double		ft_abs(double nb);
+uint32_t	get_color_from_img(mlx_image_t *img, int x, int y);
+void		texture_to_tab(t_texture *config, t_mlx *mlx);
+void		animation_sprite(t_texture *config, t_mlx *mlx);
+void		put_sprite_to_img(t_mlx *mlx, int x, t_dda *dda, t_raycast *info);
+void		put_sprite_to_img_animation(t_mlx *mlx, int x, t_dda *dda,
 			t_raycast *info);
-void	rotate_vectors(t_mlx *mlx, int flag);
-void	minimap(t_mlx *mlx);
+void		rotate_vectors(t_mlx *mlx, int flag);
+void		minimap(t_mlx *mlx);
 
 #endif
